SymbolKind workaround disabled for new versions
Microsoft has fixed the bug causing AL Language Extension to return incorrect SymbolKind value for al symbols, so workaround fixing it in AL Code Outline is no longer needed, but it should still be active if developer is using older version of Microsoft extension. To solve that problem, a few lines of code detecting if user is using AL Language Extension with that bug have been added to ALSymbolInfo class.